Sync
User
HTML

+++ src/main/java/nhom8/shoppingweb/model/User.java
@@ -10,7 +10,7 @@ import lombok.Data;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long ID;
+    private Integer ID;
     private String FULLNAME;
     private String PHONE;
     private String EMAIL;
